removed *Core files, functionality in *ParameterEditor. Better!
