+++ lib/Models/FunctionParameter.js
@@ -59,6 +59,12 @@ var FunctionParameter = function(options) {
     this.name = defaultValue(options.name, options.id);
 
     /**
+     * Gets or sets the display value of the parameter.
+     * @type {String}
+     */
+    this.displayValue = defaultValue(options.displayValue, '');
+
+    /**
      * Gets or sets the description of the parameter.
      * @type {String}
      */
