+++ lib/Models/WebProcessingServiceCatalogItem.js
@@ -172,7 +172,7 @@ WebProcessingServiceCatalogItem.prototype._load = function() {
                     return previousValue +
                         '<tr>' +
                             '<td style="vertical-align: middle">' + parameter.name + '</td>' +
-                            '<td>' + parameter.formatValueAsString(that.parameterValues[parameter.id]) + '</td>' +
+                            '<td style="padding-left: 4px">' + parameter.formatValueAsString(that.parameterValues[parameter.id]) + '</td>' +
                         '</tr>';
                 }, '') +
                 '</table>';
