+++ lib/ReactViews/Analytics/PointParameterEditor.jsx
@@ -2,9 +2,13 @@
 
 import React from 'react';
 
+import Cartographic from 'terriajs-cesium/Source/Core/Cartographic';
+import CesiumMath from 'terriajs-cesium/Source/Core/Math';
 import defined from 'terriajs-cesium/Source/Core/defined';
+import Ellipsoid from 'terriajs-cesium/Source/Core/Ellipsoid';
+import knockout from 'terriajs-cesium/Source/ThirdParty/knockout';
 
-import PointParameterEditorCore from './PointParameterEditorCore';
+import MapInteractionMode from '../../Models/MapInteractionMode';
 import ObserveModelMixin from '../ObserveModelMixin';
 
 import Styles from './parameter-editors.scss';
@@ -18,45 +22,12 @@ const PointParameterEditor = React.createClass({
         viewState: React.PropTypes.object
     },
 
-    componentWillMount() {
-        if (!defined(this.pointParameterEditorCore)) {
-            this.pointParameterEditorCore = new PointParameterEditorCore(this.props.previewed,
-                                                                         this.props.parameter,
-                                                                         this.props.viewState);
-        }
-        const value = this.pointParameterEditorCore.getValue();
-        this.setState({
-            value: value
-        });
-    },
-
-    getInitialState() {
-        let value = "";
-        if (defined(this.pointParameterEditorCore)) {
-            value = this.pointParameterEditorCore.getInitialState();
-        }
-        return {
-            value: value
-        };
-    },
-
-    onTextChange(e) {
-        this.pointParameterEditorCore.onTextChange(e);
-        this.setState({
-            value: this.pointParameterEditorCore.getValue()
-        });
-    },
-
-    getValue() {
-        return this.pointParameterEditorCore.getValue();
-    },
-
-    setValue(value) {
-        this.pointParameterEditorCore.setValue(value);
+    setDisplayValue(e) {
+        PointParameterEditor.setDisplayValue(e, this.props.parameter);
     },
 
     selectPointOnMap() {
-        this.pointParameterEditorCore.selectOnMap();
+        PointParameterEditor.selectOnMap(this.props.previewed.terria, this.props.viewState, this.props.parameter);
     },
 
     render() {
@@ -64,8 +35,8 @@ const PointParameterEditor = React.createClass({
             <div>
                 <input className={Styles.field}
                        type="text"
-                       onChange={this.onTextChange}
-                       value={this.state.value}/>
+                       onChange={this.setDisplayValue}
+                       value={this.props.parameter.displayValue}/>
                 <button type="button" onClick={this.selectPointOnMap} className={Styles.btnSelector}>
                     Select location
                 </button>
@@ -73,5 +44,97 @@ const PointParameterEditor = React.createClass({
         );
     }
 });
+
+/**
+ * Triggered when user types value directly into field.
+ * @param {String} e Text that user has entered manually.
+ * @param {FunctionParameter} parameter Parameter to set value on.
+ */
+PointParameterEditor.setDisplayValue = function(e, parameter) {
+    const coordinates = e.target.value.split(',');
+    if (coordinates.length >= 2) {
+        parameter.value = Cartographic.fromDegrees(parseFloat(coordinates[0]), parseFloat(coordinates[1]));
+    }
+};
+
+/**
+ * Given a value, return it in human readable form for display.
+ * @param {Object} value Native format of parameter value.
+ * @return {String} String for display
+ */
+PointParameterEditor.getDisplayValue = function(value) {
+    if (defined(value)) {
+        return CesiumMath.toDegrees(value.longitude) + ',' + CesiumMath.toDegrees(value.latitude);
+    } else {
+        return '';
+    }
+};
+
+/**
+ * Process value so that it can be used in an URL.
+ * @param {String} value Value to use to format.
+ * @param {FunctionParameter} parameter Parameter value belongs to.
+ * @return {String} Stringified JSON that can be used to pass parameter value in URL.
+ */
+PointParameterEditor.formatValueForUrl = function(value, parameter) {
+    if (!defined(value) || value === '') {
+        return undefined;
+    }
+
+    const coordinates = [
+        CesiumMath.toDegrees(value.longitude),
+        CesiumMath.toDegrees(value.latitude),
+    ];
+
+    if (defined(value.height)) {
+        coordinates.push(value.height);
+    }
+
+    return parameter.id + '=' + JSON.stringify({
+        'type': 'FeatureCollection',
+        'features': [
+            {
+                'type': 'Feature',
+                'geometry': {
+                    'type': 'Point',
+                    'coordinates': coordinates
+                }
+            }
+        ]
+    });
+};
+
+/**
+ * Prompt user to select/draw on map in order to define parameter.
+ * @param {Terria} terria Terria instance.
+ * @param {Object} viewState ViewState.
+ * @param {FunctionParameter} parameter Parameter.
+ */
+PointParameterEditor.selectOnMap = function(terria, viewState, parameter) {
+    // Cancel any feature picking already in progress.
+    terria.pickedFeatures = undefined;
+
+    const pickPointMode = new MapInteractionMode({
+        message: 'Select a point by clicking on the map.',
+        onCancel: function () {
+            terria.mapInteractionModeStack.pop();
+            viewState.openAddData();
+        }
+    });
+    terria.mapInteractionModeStack.push(pickPointMode);
+
+    knockout.getObservable(pickPointMode, 'pickedFeatures').subscribe(function(pickedFeatures) {
+        if (defined(pickedFeatures.pickPosition)) {
+            const value = Ellipsoid.WGS84.cartesianToCartographic(pickedFeatures.pickPosition);
+            terria.mapInteractionModeStack.pop();
+            parameter.value = value;
+            parameter.displayValue = PointParameterEditor.getDisplayValue(value);
+            parameter.processedValue = PointParameterEditor.formatValueForUrl(value, parameter);
+            viewState.openAddData();
+        }
+    });
+
+    viewState.explorerPanelIsVisible = false;
+};
 
 module.exports = PointParameterEditor;
