+++ lib/ReactViews/Analytics/RegionPicker.jsx
@@ -62,6 +62,7 @@ const RegionPicker = React.createClass({
                 }
                 this.props.parameter.value = value;
                 this.props.parameter.displayValue = this.getDisplayValue("");
+                this.props.parameter.processedValue = this.formatValueForUrl(value);
                 this.setState({
                     displayValue: this.getDisplayValue("")
                 });
@@ -97,6 +98,22 @@ const RegionPicker = React.createClass({
             this._selectedRegionCatalogItem = undefined;
         }
 
+    },
+
+    formatValueForUrl(value) {
+        const that = this;
+        return this.regionProvider.getRegionFeature(this.props.previewed.terria, value, undefined).then(function(feature) {
+            const regionParameterString = that.props.parameter.id + '=' + JSON.stringify({
+                'type': 'FeatureCollection',
+                'features': [
+                    {
+                        'type': 'Feature',
+                        'geometry': feature.geometry
+                    }
+                ]
+            });
+            return regionParameterString;
+        });
     },
 
     updateFeature(feature) {
